+++ example/example.dart
@@ -104,8 +104,7 @@ void main(List<String> args) {
     }
 
     sw.stop();
-    print(
-        'Expression "$e" parsed and and evaluated $count times in ${sw.elapsed}');
+    print('Expression "$e" parsed and evaluated $count times in ${sw.elapsed}');
   }
 
   {
@@ -125,8 +124,7 @@ void main(List<String> args) {
     }
 
     sw.stop();
-    print(
-        'Expression "$e" parsed and and evaluated $count times in ${sw.elapsed}');
+    print('Expression "$e" parsed and evaluated $count times in ${sw.elapsed}');
   }
 }