+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.4
+
+- Changes in the file `example/example.dart`.
+
 ## 0.1.3
 
 - Added support for parsing boolean literals.